refactor: 클릭 동작을 props로 받도록 수정
