+++ FE/src/components/SummaryCard/SummaryCard.jsx
@@ -1,14 +1,11 @@
 import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import moreIcon from '../../../assets/moreIcon.svg';
+import moreIcon from '../../assets/moreIcon.svg';
 import './SummaryCard.css';
-import { getDiffDays } from '../components/DateUtils';
+import { getDiffDays } from '../../pages/DashboardPage/components/DateUtils';
 
-function SummaryCard({ title, count, items = [], showDot }) {
+function SummaryCard({ title, count, items = [], showDot, onItemClick }) {
   const [isExpanded, setIsExpanded] = useState(() => window.innerWidth > 768);
   const [visibleCount, setVisibleCount] = useState(3);
-  
-  const navigate = useNavigate();
 
   //카드 헤더 클릭 시 리스트를 접거나 펼쳐짐, 접을 때 표시 개수를 초기값으로 리셋
   const toggleExpand = () => {
@@ -53,7 +50,7 @@ function SummaryCard({ title, count, items = [], showDot }) {
               <li 
                 key={item.id} 
                 className="summary-item"
-                onClick={() => navigate(`/project/${item.id}`)}
+                onClick={() => onItemClick?.(item)}
               >
                 {showDot && (
                   <span
